Fix snapshot discrepancy with travis-ci
This commit fixes the snapshot discrepancy between the reopsitory's travis-ci
and my local output.

The issue was due some platform differences as I was unable to get the expected
result on Windows 10 with `yarn locale-snapshots` command but when I executed on
WSL it emitted a different snapshot. Hope this will fix the issue.