Move argument length requirement checks to a shared function to dedupe code.
I noticed in a compiled bundle that there were a lot of duplicated strings which were used for argument length validation, so I was interested to see what effect deduplicating these would have on bundle size.

Bundle | Bytes
-------|--------
Before | `25474`
Ater   | `24528`

That's a reduction of 946 bytes (3.71%) from a complete minified bundle.