Fix Estonian date patterns
Months jaanuar and detsember were not being matched because of
the leading and trailing single quotes (') in the pattern.

Days of week pühapäev and laupäev were not being matched because
of the leading and trailing single quotes (') in the pattern.

Months juuni and juuli were not being matched because of the
missing double "u" in the pattern.

Fixes #1571.