Add "1" to time periods
Standalone period names are only usable in a sentence, but not in standalone text. Also their English locale equivalent also states the number 1 in front of the period name.

Adding 1 to the period name does not break current usage, but adds usability.

Consider following Czech sentence, which retains same meaning, possibly adding more readability:
Zbývá {{distance}} – Zbývá hodina 40 minut. -> Zbývá 1 hodina 40 minut.

Consider following GUI, which currently is not usable and gains usability:
⏱️ {{distance}} – ⏱️ hodina 40 minut -> ⏱️ 1 hodina 40 minut

That is the reason to include the extra 1 in front of the period name.