Optionally omit timezone offset from `formatISO`
Adds an option to `formatISO`, `includeOffset`, which can be set to
`false` to omit the timezone offset from the formatted string. This
option defaults to `true`, maintaining the existing behavior unless it
is explicitly set to `false` by a consumer.

The option is implemented in `FormatISOOptions` instead of the shared
`ISOFormatOptions` interface because the latter is shared by
`formatISO99075`, which does not support timezone offsets.